Workaround bug in mgo Find() method.
The bug occurs when trying to Find(bson.M{"_id", <id>}) (or use the relevant FindId() method). Instead of getting the object, it simply returns your requested struct, with the only field filled being the ID. It should do reflection and retrieve struct member names like it does with Create(), but it doesn't.

Looks like I need to file this - confirmed this bug during the hackathon w/ MongoDB dev person.

Signed-off-by: Robert Xu <robxu9@gmail.com>